python/requirements.txt
Pillow>=10.1.0
blurhash-python>=1.2.1